requirements.txt
aiohttp==3.8.1; python_version >= "3.6" and python_full_version >= "3.8.0"
aiosignal==1.2.0; python_version >= "3.6" and python_full_version >= "3.8.0"
async-timeout==4.0.2; python_version >= "3.6" and python_full_version >= "3.8.0"
attrs==22.1.0; python_version >= "3.6" and python_full_version >= "3.8.0"
charset-normalizer==2.1.1; python_version >= "3.6" and python_full_version >= "3.8.0"
discord.py==2.0.0; python_full_version >= "3.8.0"
discord==2.0.0
frozenlist==1.3.1; python_version >= "3.7" and python_full_version >= "3.8.0"
idna==3.3; python_version >= "3.7" and python_full_version >= "3.8.0"
multidict==6.0.2; python_version >= "3.7" and python_full_version >= "3.8.0"
python-dotenv==0.20.0; python_version >= "3.5"
yarl==1.8.1; python_version >= "3.7" and python_full_version >= "3.8.0"
